// frontend/src/app/core/api/model/student-enrollment-simple-response.ts
/**
 * Training school system Backend API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { PromotionSummaryResponse } from './promotion-summary-response';


export interface StudentEnrollmentSimpleResponse { 
    id?: number;
    enrollmentNumber?: string;
    type?: StudentEnrollmentSimpleResponse.TypeEnum;
    status?: StudentEnrollmentSimpleResponse.StatusEnum;
    promotion?: PromotionSummaryResponse;
}
export namespace StudentEnrollmentSimpleResponse {
    export const TypeEnum = {
        New: 'NEW',
        Repeater: 'REPEATER',
        Reintegration: 'REINTEGRATION'
    } as const;
    export type TypeEnum = typeof TypeEnum[keyof typeof TypeEnum];
    export const StatusEnum = {
        PreEnrolled: 'PRE_ENROLLED',
        Incomplete: 'INCOMPLETE',
        Waitlisted: 'WAITLISTED',
        ConditionallyValidated: 'CONDITIONALLY_VALIDATED',
        Validated: 'VALIDATED',
        Suspended: 'SUSPENDED',
        DroppedOut: 'DROPPED_OUT',
        Rejected: 'REJECTED',
        Cancelled: 'CANCELLED',
        Completed: 'COMPLETED'
    } as const;
    export type StatusEnum = typeof StatusEnum[keyof typeof StatusEnum];
}